requirements.txt
streamlit
torch
timm
pillow
